rails/image_to_text_server/requirements.txt
fastapi==0.115.3
uvicorn==0.32.0
requests==2.32.3
numpy
matplotlib
accelerate==0.25.0
huggingface-hub==0.20.1
Pillow==10.1.0
torch==2.1.2
torchvision==0.16.2
transformers==4.36.2
einops==0.7.0
sentence-transformers==2.2.2
